move public records to native
