+++ README.md
@@ -19,6 +19,7 @@
 ### 创建和使用客户端
 
 公共接口：
+
 ```
 ai.bianjie.avatasdk.AvataClient client = new ai.bianjie.avatasdk.AvataClient.Builder()
             .setDoMain("")
@@ -29,6 +30,7 @@ ai.bianjie.avatasdk.AvataClient client = new ai.bianjie.avatasdk.AvataClient.Bui
 ```
 
 原生接口：
+
 ```
 ai.bianjie.avatasdk.AvataNativeClient client = new ai.bianjie.avatasdk.AvataNativeClient.Builder()
             .setDomain("")
@@ -39,8 +41,8 @@ ai.bianjie.avatasdk.AvataNativeClient client = new ai.bianjie.avatasdk.AvataNati
 
 ```
 
-
 智能合约接口：
+
 ```
 ai.bianjie.avatasdk.AvataEvmClient client = new ai.bianjie.avatasdk.AvataEvmClient.Builder()
             .setDomain("")
@@ -51,17 +53,16 @@ ai.bianjie.avatasdk.AvataEvmClient client = new ai.bianjie.avatasdk.AvataEvmClie
 
 ```
 
-
 公共接口：
 
 - [链账户接口示例代码](./src/test/java/AccountTest.java)
 - [充值接口示例代码](./src/test/java/OrderTest.java)
-- [存证接口示例代码](./src/test/java/RecordsTest.java)
 
 原生接口：
 
 - [NFT 接口示例代码](./src/test/java/nat/NftTest.java)
 - [MT 接口示例代码](./src/test/java/nat/MtTest.java)
+- [存证接口示例代码](./src/test/java/nat/RecordsTest.java)
 - [交易结果查询接口示例代码](./src/test/java/nat/TxTest.java)
 
 智能合约接口：
