+++ src/main/java/ai/bianjie/avatasdk/AvataClient.java
@@ -4,7 +4,7 @@ import ai.bianjie.avatasdk.config.ConfigInfo;
 import ai.bianjie.avatasdk.exception.AvataException;
 import ai.bianjie.avatasdk.proxy.account.impl.AccountClient;
 import ai.bianjie.avatasdk.proxy.order.impl.OrderClient;
-import ai.bianjie.avatasdk.proxy.records.impl.RecordsClient;
+import ai.bianjie.avatasdk.proxy.nat.records.impl.RecordsClient;
 import ai.bianjie.avatasdk.proxy.user.impl.UserClient;
 import com.dtflys.forest.Forest;
 import com.dtflys.forest.config.ForestConfiguration;
@@ -17,7 +17,6 @@ public class AvataClient {
 
     public AccountClient accountClient;
     public OrderClient orderClient;
-    public RecordsClient recordsClient;
     public UserClient userClient;
 
     /**
@@ -35,7 +34,6 @@ public class AvataClient {
 
         this.orderClient = new OrderClient(configInfo);
 
-        this.recordsClient = new RecordsClient(configInfo);
         this.userClient = new UserClient(configInfo);
 
     }
